requirements.txt
absl-py                    2.3.0
accelerate                 1.1.1
aioboto3                   13.2.0
aiobotocore                2.15.2
aiofiles                   24.1.0
aiohappyeyeballs           2.4.3
aiohttp                    3.11.2
aioitertools               0.12.0
aiosignal                  1.3.1
alembic                    1.14.0
annotated-types            0.7.0
anyio                      4.6.2.post1
anytree                    2.12.1
asttokens                  3.0.0
attrs                      24.2.0
autograd                   1.7.0
banal                      1.0.6
beartype                   0.18.5
beautifulsoup4             4.12.3
blis                       1.0.1
boto3                      1.35.36
botocore                   1.35.36
Brotli                     1.0.9
catalogue                  2.0.10
certifi                    2024.8.30
cffi                       1.17.1
charset-normalizer         3.3.2
click                      8.1.7
cloudpathlib               0.20.0
colorama                   0.4.6
concepcy                   0.1.0
confection                 0.1.5
contourpy                  1.3.1
cryptography               44.0.0
cssselect                  1.2.0
curated-tokenizers         0.0.9
curated-transformers       0.1.1
cycler                     0.12.1
cymem                      2.0.10
dashscope                  1.20.13
dataset                    1.6.2
datasets                   3.1.0
decorator                  5.1.1
dill                       0.3.8
distro                     1.9.0
docker-pycreds             0.4.0
easydict                   1.13
en_core_web_trf            3.8.0
evaluate                   0.4.3
executing                  2.1.0
faiss                      1.9.0
filelock                   3.13.1
fonttools                  4.55.0
frozenlist                 1.5.0
fsspec                     2024.9.0
ftfy                       6.3.1
gensim                     4.3.3
geomloss                   0.2.6
gitdb                      4.0.11
GitPython                  3.1.43
graspologic                3.4.1
graspologic-native         1.2.1
greenlet                   3.1.1
h11                        0.14.0
hnswlib                    0.8.0
httpcore                   1.0.7
httpx                      0.27.2
huggingface-hub            0.26.2
hyppo                      0.4.0
idna                       3.10
inquirerpy                 0.3.4
ipython                    8.30.0
jedi                       0.19.2
Jinja2                     3.1.4
jiter                      0.7.1
jmespath                   1.0.1
joblib                     1.4.2
jsonpickle                 4.0.0
kiwisolver                 1.4.7
langcodes                  3.5.0
language_data              1.3.0
lightning-utilities        0.11.9
lightrag-hku               1.0.2
llvmlite                   0.43.0
loralib                    0.1.2
lxml                       5.3.0
Mako                       1.3.6
marisa-trie                1.2.1
markdown-it-py             3.0.0
MarkupSafe                 2.1.3
matplotlib                 3.9.3
matplotlib-inline          0.1.7
mdurl                      0.1.2
mkl_fft                    1.3.11
mkl_random                 1.2.8
mkl-service                2.4.0
mpmath                     1.3.0
multidict                  6.1.0
multiprocess               0.70.16
murmurhash                 1.0.11
nano-vectordb              0.0.4.3
neo4j                      5.27.0
networkx                   3.2.1
nltk                       3.9.1
numba                      0.60.0
numpy                      1.26.4
ollama                     0.4.2
openai                     1.54.4
oracledb                   2.5.0
packaging                  24.1
pandas                     2.2.3
parso                      0.8.4
patsy                      1.0.1
peft                       0.14.0
pfzy                       0.3.4
pillow                     11.0.0
pip                        24.2
platformdirs               4.3.6
POT                        0.9.5
preshed                    3.0.9
prompt_toolkit             3.0.48
propcache                  0.2.0
protobuf                   5.29.0
psutil                     6.1.0
pure_eval                  0.2.3
pyarrow                    18.0.0
pycparser                  2.22
pydantic                   2.10.2
pydantic_core              2.27.1
Pygments                   2.18.0
pynndescent                0.5.13
pyparsing                  3.2.0
pyquery                    2.0.1
PySocks                    1.7.1
python-dateutil            2.9.0.post0
pytz                       2024.2
pyvis                      0.3.2
PyYAML                     6.0.2
rank-bm25                  0.2.2
regex                      2024.11.6
request-boost              0.6
requests                   2.32.3
rich                       13.9.4
rouge                      1.0.1
rouge_score                0.1.2
s3transfer                 0.10.4
safetensors                0.4.5
scikit-learn               1.5.2
scipy                      1.12.0
seaborn                    0.13.2
sentence-transformers      3.3.1
sentencepiece              0.2.0
sentry-sdk                 2.19.2
setproctitle               1.3.4
setuptools                 75.1.0
shellingham                1.5.4
six                        1.16.0
smart-open                 7.0.5
smmap                      5.0.1
sniffio                    1.3.1
soupsieve                  2.6
spacy                      3.8.2
spacy-curated-transformers 0.3.0
spacy-legacy               3.0.12
spacy-loggers              1.0.5
SQLAlchemy                 1.4.54
srsly                      2.4.8
stack-data                 0.6.3
statsmodels                0.14.4
sympy                      1.13.1
tenacity                   9.0.0
thinc                      8.3.2
threadpoolctl              3.5.0
tiktoken                   0.8.0
tokenizers                 0.20.3
torch                      2.5.1
torchaudio                 2.5.1
torchmetrics               1.6.0
torchvision                0.20.1
tqdm                       4.67.0
traitlets                  5.14.3
transformers               4.46.3
typer                      0.13.1
typing_extensions          4.12.2
tzdata                     2024.2
umap-learn                 0.5.7
urllib3                    2.2.3
wandb                      0.19.1
wasabi                     1.1.3
wcwidth                    0.2.13
weasel                     0.4.1
websocket-client           1.8.0
wheel                      0.44.0
wikipedia                  1.4.0
win-inet-pton              1.1.0
wrapt                      1.17.0
xxhash                     3.5.0
yarl                       1.17.2